requirements.txt
python-telegram-bot
pandas
gspread
emoji
